requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
asyncstdlib==3.13.3
    # via denonavr
attrs==25.4.0
    # via denonavr
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
defusedxml==0.7.1
    # via denonavr
denonavr==1.3.2
    # via denon-proxy (pyproject.toml)
ftfy==6.3.1
    # via denonavr
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   denon-proxy (pyproject.toml)
    #   denonavr
idna==3.11
    # via
    #   anyio
    #   httpx
ifaddr==0.2.0
    # via zeroconf
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
netifaces==0.11.0
    # via denonavr
pydantic==2.12.5
    # via denon-proxy (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyyaml==6.0.3
    # via denon-proxy (pyproject.toml)
rich==14.3.3
    # via denon-proxy (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
wcwidth==0.6.0
    # via ftfy
zeroconf==0.148.0
    # via denon-proxy (pyproject.toml)
